refactor(cli): organize token commands
