+++ app/src/cli/commands/token/mod.rs
@@ -1,3 +1,6 @@
+mod args;
 mod handler;
 
+pub(crate) use args::HELP;
+pub use args::TokenCommand;
 pub(super) use handler::execute;
